Consolidate project list component
